carrito funcional con calculo de total de compra
