+++ client/src/clientComponents/ShoppingCar/Checkout.jsx
@@ -1,7 +1,6 @@
-import React, { useEffect } from "react";
+import React, { useEffect, useState } from "react";
 import classnames from 'classnames'
 import { Context } from "./ContextProvider";
-import Item from './Item.jsx'
 
 const Checkout = ({ onClick }) => {
   const [isVisible, setIsVisible] = React.useState(true);
@@ -10,17 +9,32 @@ const Checkout = ({ onClick }) => {
     'shopping-cart--hidden': !isVisible,
   })
   
+  const updatePrice = (event) => {
+    const quantity = event.target.value;
+    const variation=quantity-orderData.quantity;
+    const name=event.target.name;
+    const item=orderData.filter(it=>it.name===name)[0]
+    const amount= parseInt(item.final_price) * parseInt(quantity);
+    console.log(amount)
+    const actual=[]
+    orderData.map(it=>{if(it.name===name){
+        actual.push({'id':it.id,'name':it.name,'image':it.image,'final_price':it.final_price,quantity:parseInt(quantity),amount:amount})
+      }else{actual.push(it)}
+    })
+    setOrderData(actual)
+    console.log(orderData)
+    total=total+variation*orderData.final_price;
+  }
 
   useEffect(() => {
     if (preferenceId) setIsVisible(false);
   }, [preferenceId])
 
+let total=0
+orderData.forEach(item=>{
+  total=total+item.amount;
+})
 
-  // const updatePrice = (event) => {
-  //   const quantity = event.target.value;
-  //   const amount = parseInt(orderData.finaly_price) * parseInt(quantity);
-  //   setOrderData({ ...orderData, quantity, amount });
-  // }
   
   return (
     <section className={shoppingCartClass}>
@@ -70,10 +84,54 @@ const Checkout = ({ onClick }) => {
 
           {orderData.map(item=>{
             return(
-              <Item
-                orderData = {orderData}
-                name={item.name}
-              />
+              
+              <div className="row">
+      <div className="col-md-12 col-lg-8">
+        <div className="items">
+          <div className="product">
+            <div className="info">
+              <div className="product-details">
+                <div className="row justify-content-md-center">
+                  <div className="col-md-3">
+                    <img
+                      className="img-fluid mx-auto d-block image"
+                      alt="Image of a product"
+                      src={item.image}
+                    />
+                  </div>
+                  <div className="col-md-4 product-detail">
+                    <div className="product-info">
+                      <b>{item.name}</b><br></br>
+                      <b>Price:</b> $ <span id="unit-price">{item.final_price}</span>
+                      <br />
+                    </div>
+                  </div>
+                  <div className="col-md-3 product-detail">
+                    <input
+                      onChange={updatePrice}
+                      type="number"
+                      id="quantity"
+                      name={item.name}
+                      value={item.quantity}
+                      min="1"
+                      className="form-control"
+                    />
+                  </div>
+                </div>
+              </div>
+            </div>
+          </div>
+        </div>
+      </div>
+      <div className="col-md-12 col-lg-4">
+        <div className="summary">
+          <div className="summary-item">
+            <span className="price" id="cart-total">${item.amount}</span>
+          </div>
+        </div>
+      </div>
+    </div>
+
             )
           })}
 
@@ -102,7 +160,8 @@ const Checkout = ({ onClick }) => {
             <div className="col-md-12 col-lg-4">
               <div className="summary">
                 <div className="summary-item">
-                  {/*espacio para el subtotal*/}
+                  <span className="text">Total</span>
+                  <span className="price" id="cart-total">${total}</span>
                 </div>
                 <button
                   className="btn btn-primary btn-lg btn-block"
