+++ client/src/redux/foodSlice.js
@@ -11,24 +11,7 @@ const initialState = {
   allFoods: [],
   foodDetail: [],
   filteredFoods:[],
-  orderItems:[
-    {
-      final_price : 3200,
-      id : "85017bae-4e74-4e1b-9092-1d39a7d33d34",
-      image : "https://assets.kraftfoods.com/recipe_images/opendeploy/94128_640x428.jpg",
-      name : "Bistec a la parrilla",
-      quantity: 1,
-      amount: 3200,
-    },
-    {
-      final_price : 2800,
-      id : "9cdc2214-7fc4-44b4-8bcb-c46c9ac15acd",
-      image : "https://danzadefogones.com/wp-content/uploads/2018/03/Pasta-primavera-2.jpg",
-      name : "Pasta Primavera",
-      quantity: 1,
-      amount: 2800,
-    },
-  ],
+  orderItems:[],
   activeFilteredFoods:false,
   orderBy:'',
   foodsCategory:'',
